requirements.txt
Pillow
-U langchain-community
langchain-google-genai
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
crewai  
SpeechRecognition
